InMemoryRecordStorage supports similarity search.
